完善增删改查

+++ src/main/resources/sql/init.sql
@@ -10,7 +10,8 @@ CREATE TABLE `t_user`
     `id`     bigint(20) NOT NULL AUTO_INCREMENT,
     `name`   varchar(20) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci NULL DEFAULT NULL,
     `age`    int(3) NULL DEFAULT NULL,
-    `salary` double(7, 2) NULL DEFAULT NULL,
+    `salary` double(9, 2
+) NULL DEFAULT NULL,
   `phone` int(11) NULL DEFAULT NULL,
   PRIMARY KEY (`id`) USING BTREE
 ) ENGINE = InnoDB AUTO_INCREMENT = 1 CHARACTER SET = utf8mb4 COLLATE = utf8mb4_general_ci ROW_FORMAT = Dynamic;
